Move stuff into provider module
