+++ experimental/casemapping/src/exceptions.rs
@@ -6,7 +6,8 @@ use icu_provider::prelude::*;
 use zerovec::{VarZeroVec, ZeroVec};
 
 use crate::error::Error;
-use crate::internals::{ClosureSet, DotType, MappingKind};
+use crate::internals::ClosureSet;
+use crate::provider::data::{DotType, MappingKind};
 
 /// This represents case mapping exceptions that can't be represented as a delta applied to
 /// the original code point. Similar to ICU4C, data is stored as a u16 array. The codepoint
