+++ experimental/casemapping/src/provider.rs
@@ -13,6 +13,9 @@ pub use crate::internals::CaseMappingData;
 pub use crate::internals::CaseMappingInternals;
 pub use crate::internals::CaseMappingUnfoldData;
 
+#[allow(missing_docs)] // TBD, temporary
+pub mod data;
+
 #[icu_provider::data_struct(CaseMappingV1Marker = "props/casemap@1")]
 #[derive(Debug, PartialEq, Clone)]
 #[cfg_attr(feature = "serde", derive(serde::Deserialize))]
